Coercing to numbers, strings, and latex:
* Fraction.print -> Fraction.toString
* Fraction.tex -> Fraction.toTex
* Fraction.decimal -> Fraction.valueOf

* Number.toTex

* Expression.print -> Expression.toString
* Expression.tex -> Expression.toTex

* Equation.print -> Equation.toString
* Equation.tex -> Equation.toTex

When solving a quadratic equation, you might get an array of Numbers or an array of Fractions, which is dependent upon whether or not the answers are rational. Previously, you'd have to check the type of the object before printing to string or tex or getting the value. This is because the names of the Fraction methods were different than the names of the Number methods. But now, you can do:

Object.toString()
Object.toTex()
Object.valueOf()

regardless of whether or not the object is a Fraction or a Number. Yay!

Object.tex was changed to Object.toTex to mirror toString.