Linear equations can't have cross products.
Previously, Equation._isLinear was only checking that the maximum degree
is 1, but you also need to check that there are no cross products. E.g.,

xy = 2

is not linear. There was also a bug in
Equation._variableCanBeIsolated(variable) due to the same issue where
cross products were not being checked. This change fixes both these
issues.