Negatives weren't being stripped from Terms with Fractions
Term.toTex and Term.toString are supposed to strip any negative signs
at the beginning since that sign will be brought out front in
Expression.toTex and Expression.toString. This was failing due to a bug
in Term.toTex; the negative was not being stripped out because it was
only checking for a “-“ at the beginning of the string. This missed
cases where the coefficient is a fraction, as the negative is printed
on the numerator and not at the beginning, e.g. \\frac{-3}{4}.
Term.toTex now correctly strips out negative signs where the first
coefficient is a fraction.

Fixes #32.